add cluster autoscaller
