+++ modules/040-node-manager/hooks/enable_cluster_api_cloud_and_static.go
@@ -17,28 +17,13 @@ limitations under the License.
 package hooks
 
 import (
-	"context"
-	"os"
-	"time"
-
 	"github.com/flant/addon-operator/pkg/module_manager/go_hook"
 	"github.com/flant/addon-operator/sdk"
-	"github.com/flant/shell-operator/pkg/kube/object_patch"
 	"github.com/flant/shell-operator/pkg/kube_events_manager/types"
-	"github.com/pkg/errors"
-	certificatesv1 "k8s.io/api/certificates/v1"
 	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/deckhouse/deckhouse/go_lib/dependency"
-	"github.com/deckhouse/deckhouse/go_lib/dependency/k8s"
-	"github.com/deckhouse/deckhouse/go_lib/hooks/tls_certificate"
-	"github.com/deckhouse/deckhouse/modules/040-node-manager/hooks/internal/kubeconfig"
-
 	ngv1 "github.com/deckhouse/deckhouse/modules/040-node-manager/hooks/internal/v1"
 )
 
@@ -141,109 +126,14 @@ func handleClusterAPIDeploymentRequired(input *go_hook.HookInput, dc dependency.
 
 	if capiEnabled {
 		input.Values.Set("nodeManager.internal.capiControllerManagerEnabled", true)
-		if _, ok := os.LookupEnv("TEST_SKIP_GENERATE_KUBECONFIG"); !ok {
-			err := generateStaticKubeconfigSecret(input, dc, hookParam{
-				serviceAccount: clusterAPICloudServiceAccountName,
-				cluster:        capiClusterName,
-			})
-			if err != nil {
-				return err
-			}
-		}
 	} else {
 		input.Values.Remove("nodeManager.internal.capiControllerManagerEnabled")
 	}
 
 	if capsEnabled || hasStaticInstancesField {
-		if _, ok := os.LookupEnv("TEST_SKIP_GENERATE_KUBECONFIG"); !ok {
-			err := generateStaticKubeconfigSecret(input, dc, hookParam{
-				serviceAccount: clusterAPIStaticServiceAccountName,
-				cluster:        clusterAPIStaticClusterName,
-			})
-			if err != nil {
-				return err
-			}
-		}
-
 		input.Values.Set("nodeManager.internal.capsControllerManagerEnabled", true)
 	} else {
 		input.Values.Remove("nodeManager.internal.capsControllerManagerEnabled")
-	}
-
-	return nil
-}
-
-func generateStaticKubeconfigSecret(input *go_hook.HookInput, dc dependency.Container, params hookParam) error {
-	restConfig, err := dc.GetClientConfig()
-	if err != nil {
-		return errors.Wrap(err, "failed to get kubeconfig")
-	}
-
-	k8sClient, err := dc.GetK8sClient()
-	if err != nil {
-		return errors.Wrap(err, "failed to get k8s client")
-	}
-
-	err = createCAPIServiceAccount(k8sClient, params.serviceAccount)
-	if err != nil {
-		return errors.Wrap(err, "failed to create Cluster API service account")
-	}
-
-	certExirationSeconds := int32((180 * 24 * time.Hour).Seconds())
-
-	cert, err := tls_certificate.IssueCertificate(input, dc, tls_certificate.OrderCertificateRequest{
-		CommonName: "capi-controller-manager",
-		Groups: []string{
-			"d8:node-manager:capi-controller-manager:manager-role",
-		},
-		Usages: []certificatesv1.KeyUsage{
-			certificatesv1.UsageClientAuth,
-		},
-		ExpirationSeconds: &certExirationSeconds,
-	})
-	if err != nil {
-		return errors.Wrap(err, "failed to issue certificate")
-	}
-
-	config, err := kubeconfig.New(params.cluster, restConfig.Host, restConfig.CAData, []byte(cert.Key), []byte(cert.Certificate))
-	if err != nil {
-		return errors.Wrap(err, "failed to generate a kubeconfig")
-	}
-
-	configYAML, err := clientcmd.Write(*config)
-	if err != nil {
-		return errors.Wrap(err, "failed to serialize kubeconfig to yaml")
-	}
-
-	secret := kubeconfig.GenerateSecret(params.cluster, clusterAPINamespace, configYAML)
-
-	secretUnstructured, err := runtime.DefaultUnstructuredConverter.ToUnstructured(secret)
-	if err != nil {
-		return errors.Wrap(err, "failed to convert secret to unstructured")
-	}
-
-	input.PatchCollector.Create(secretUnstructured, object_patch.UpdateIfExists())
-
-	return nil
-}
-
-func createCAPIServiceAccount(k8sClient k8s.Client, saName string) error {
-	serviceAccount := &corev1.ServiceAccount{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "v1",
-			Kind:       "ServiceAccount",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      saName,
-			Namespace: clusterAPINamespace,
-		},
-	}
-
-	_, err := k8sClient.CoreV1().ServiceAccounts(serviceAccount.Namespace).Create(context.TODO(), serviceAccount, metav1.CreateOptions{})
-	if err != nil {
-		if !apierrors.IsAlreadyExists(err) {
-			return errors.Wrap(err, "failed to create service account")
-		}
 	}
 
 	return nil
